+++ src/main/java/com/dantonov/musicstore/service/ActionService.java
@@ -18,7 +18,7 @@ public class ActionService {
     
     
     
-    public Action findById(Long id) {
+    public Action findById(Byte id) {
         return actionRepository.findOne(id);
     }
